Move Hierarchy to Type
