+++ src/Check/CheckIfString.php
@@ -12,21 +12,12 @@ namespace Vivarium\Check;
 
 use Vivarium\Assertion\String\Contains;
 use Vivarium\Assertion\String\EndsWith;
-use Vivarium\Assertion\Type\IsBasicType;
-use Vivarium\Assertion\Type\IsClass;
-use Vivarium\Assertion\Type\IsClassOrInterface;
 use Vivarium\Assertion\String\IsEmpty;
-use Vivarium\Assertion\Type\IsInterface;
-use Vivarium\Assertion\Type\IsIntersection;
 use Vivarium\Assertion\String\IsLong;
 use Vivarium\Assertion\String\IsLongAtLeast;
 use Vivarium\Assertion\String\IsLongAtMax;
 use Vivarium\Assertion\String\IsLongBetween;
-use Vivarium\Assertion\String\IsNamespace;
 use Vivarium\Assertion\String\IsNotEmpty;
-use Vivarium\Assertion\Type\IsPrimitive;
-use Vivarium\Assertion\Type\IsType;
-use Vivarium\Assertion\Type\IsUnion;
 use Vivarium\Assertion\String\StartsWith;
 
 final class CheckIfString
@@ -41,38 +32,9 @@ final class CheckIfString
         return (new EndsWith($end))($string);
     }
 
-    /** @psalm-assert-if-true 'int'|'float'|'string'|'array'|'callable'|'object'|class-string $string */
-    public static function isBasicType(string $string): bool
-    {
-        return (new IsBasicType())($string);
-    }
-
-    /** @psalm-assert-if-true class-string $string */
-    public static function isClass(string $string): bool
-    {
-        return (new IsClass())($string);
-    }
-
-    /** @psalm-assert-if-true class-string $string */
-    public static function isClassOrInterface(string $string): bool
-    {
-        return (new IsClassOrInterface())($string);
-    }
-
     public static function isEmpty(string $string): bool
     {
         return (new IsEmpty())($string);
-    }
-
-    /** @psalm-assert-if-true class-string $string */
-    public static function isInterface(string $string): bool
-    {
-        return (new IsInterface())($string);
-    }
-
-    public static function isIntersection(string $string): bool
-    {
-        return (new IsIntersection())($string);
     }
 
     public static function isLong(string $string, int $length, string $encoding = 'UTF-8'): bool
@@ -95,29 +57,9 @@ final class CheckIfString
         return (new IsLongBetween($min, $max, $encoding))($string);
     }
 
-    public static function isNamespace(string $string): bool
-    {
-        return (new IsNamespace())($string);
-    }
-
     public static function isNotEmpty(string $string): bool
     {
         return (new IsNotEmpty())($string);
-    }
-
-    public static function isPrimitive(string $string): bool
-    {
-        return (new IsPrimitive())($string);
-    }
-
-    public static function isType(string $string): bool
-    {
-        return (new IsType())($string);
-    }
-
-    public static function isUnion(string $string): bool
-    {
-        return (new IsUnion())($string);
     }
 
     public static function startsWith(string $string, string $start): bool
